session from everywhere

+++ types/player.go
@@ -1,9 +1,8 @@
 package types
 
-
 type Player struct {
-	X 		int
-	Y 		int
-	Blood	int
-	Score 	int
+	X     float64
+	Y     float64
+	Blood int
+	Score int
 }
